# backend/core-api/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_AvinashMahala_ClusterGenie_backend_core-api_models.ClusterResponse:
    properties:
      cluster:
        $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.Cluster'
      message:
        type: string
    type: object
  github_com_AvinashMahala_ClusterGenie_backend_core-api_models.CreateClusterRequest:
    properties:
      name:
        type: string
      region:
        type: string
    type: object
...
  github_com_AvinashMahala_ClusterGenie_backend_core-api_models.DeleteClusterResponse:
    properties:
      message:
        type: string
    type: object
...
  github_com_AvinashMahala_ClusterGenie_backend_core-api_models.DiagnoseClusterRequest:
    properties:
      cluster_id:
        type: string
    type: object
  github_com_AvinashMahala_ClusterGenie_backend_core-api_models.DiagnoseClusterResponse:
    properties:
      cluster:
        $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.Cluster'
      insights:
        items:
          type: string
        type: array
      recommendations:
        items:
          type: string
        type: array
    type: object
...
  github_com_AvinashMahala_ClusterGenie_backend_core-api_models.ErrorResponse:
    properties:
      error:
        description: |-
          The error message
          example: Invalid request payload
        type: string
    type: object
...
  github_com_AvinashMahala_ClusterGenie_backend_core-api_models.ListClustersResponse:
    properties:
      clusters:
        items:
          $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.Cluster'
        type: array
    type: object
...
  github_com_AvinashMahala_ClusterGenie_backend_core-api_models.UpdateClusterRequest:
    properties:
      name:
        type: string
      region:
        type: string
      status:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: REST API for ClusterGenie - DigitalOcean droplet management and monitoring
  title: ClusterGenie API
  version: "1.0"
paths:
  /clusters:
    get:
      consumes:
      - application/json
      description: Returns a list of clusters
      produces:
      - application/json
      responses:
        "200":
          description: List of clusters
          schema:
            $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.ListClustersResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.ErrorResponse'
      summary: List clusters
      tags:
      - clusters
    post:
      consumes:
      - application/json
      description: Create a new cluster entity
      parameters:
      - description: Create cluster request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.CreateClusterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Cluster created
          schema:
            $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.ClusterResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.ErrorResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.ErrorResponse'
      summary: Create a new cluster
      tags:
      - clusters
  /clusters/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes the cluster and returns a confirmation message
      parameters:
      - description: Cluster ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Deletion confirmation
          schema:
            $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.DeleteClusterResponse'
        "404":
          description: Cluster not found
          schema:
            $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.ErrorResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.ErrorResponse'
      summary: Delete cluster
      tags:
      - clusters
    get:
      consumes:
      - application/json
      description: Retrieve a cluster by its id
      parameters:
      - description: Cluster ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Cluster retrieved
          schema:
            $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.ClusterResponse'
        "404":
          description: Cluster not found
          schema:
            $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.ErrorResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.ErrorResponse'
      summary: Get cluster by id
      tags:
      - clusters
    put:
      consumes:
      - application/json
      description: Update an existing cluster's fields
      parameters:
      - description: Cluster ID
        in: path
        name: id
        required: true
        type: string
      - description: Update cluster request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.UpdateClusterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Updated cluster
          schema:
            $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.ClusterResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.ErrorResponse'
        "404":
          description: Cluster not found
          schema:
            $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.ErrorResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.ErrorResponse'
      summary: Update cluster
      tags:
      - clusters
  /diagnosis/diagnose:
    post:
      consumes:
      - application/json
      description: Run a diagnosis on the supplied cluster and get insights/recommendations
      parameters:
      - description: Diagnose request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.DiagnoseClusterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Diagnosis result
          schema:
            $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.DiagnoseClusterResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.ErrorResponse'
        "429":
          description: Rate limit exceeded
          schema:
            $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.ErrorResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.ErrorResponse'
      summary: Diagnose cluster
      tags:
      - diagnosis
  /droplets:
    get:
      consumes:
      - application/json
      description: Returns a list of droplets (optionally filtered by cluster)
      produces:
      - application/json
      responses:
        "200":
          description: List of droplets
          schema:
            $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.ListDropletsResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.ErrorResponse'
      summary: List droplets
      tags:
      - droplets
    post:
      consumes:
      - application/json
      description: Provisions a new droplet. Provide region/size/image and optionally
        a cluster_id
      parameters:
      - description: Create droplet request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.CreateDropletRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Droplet created
          schema:
            $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.DropletResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.ErrorResponse'
        "500":
          description: Server error while provisioning
          schema:
            $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.ErrorResponse'
      summary: Create a new droplet
      tags:
      - droplets
  /droplets/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a droplet and returns a confirmation message
      parameters:
      - description: Droplet ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Deletion confirmation
          schema:
            $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.DeleteDropletResponse'
        "404":
          description: Droplet not found
          schema:
            $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.ErrorResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.ErrorResponse'
      summary: Delete droplet
      tags:
      - droplets
    get:
      consumes:
      - application/json
      description: Returns a single droplet by id
      parameters:
      - description: Droplet ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Droplet retrieved
          schema:
            $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.DropletResponse'
        "404":
          description: Droplet not found
          schema:
            $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.ErrorResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.ErrorResponse'
      summary: Get droplet by id
      tags:
      - droplets
...
  /hello:
    post:
      consumes:
      - application/json
      description: Returns a greeting message for the supplied name
      parameters:
      - description: Hello request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.HelloRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Greeting created
          schema:
            $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.HelloResponse'
        "400":
          description: Invalid request payload
          schema:
            $ref: '#/definitions/github_com_AvinashMahala_ClusterGenie_backend_core-api_models.ErrorResponse'
      summary: Say hello
      tags:
      - hello
...
swagger: "2.0"